Outputs Base64 over multiple lines
It is unclear if this will fix GitHub issues rendering this content but
it should make it nicer on text editors like sublime that struggle with
lines too large in width.

Various extraneous pieces of code removed here too. Borne out of fatigue
most likely!

The screenshot dimensions are now confiurable in JSON which should be a
nice optimization meaning we won't need to recompile so often.